requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.6.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.3.1
Django==2.0.4
django-modelcluster==4.1
django-taggit==0.22.2
django-treebeard==4.3
djangorestframework==3.8.2
draftjs-exporter==2.1.2
Flask==1.0.2
gunicorn==19.9.0
html5lib==1.0.1
idna==2.7
itsdangerous==0.24
Jinja2==2.10
kafka-helper==0.2
kafka-python==1.4.3
l18n==2016.6.4
MarkupSafe==1.0
nose==1.3.7
numpy==1.14.2
pika==0.12.0
Pillow==5.1.0
pycparser==2.19
python-dotenv==0.9.1
pytz==2018.4
requests==2.19.1
six==1.11.0
Unidecode==0.4.21
urllib3==1.23
wagtail==2.1
webencodings==0.5.1
Werkzeug==0.14.1
Willow==1.1
